requirements.txt

python-dotenv==1.0.0
random2==1.0.1
openai==0.27.8
gtts==2.2.3
moviepy==1.0.3
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6